+++ NEWS.md
@@ -3,3 +3,11 @@
 ## Version 0.1.0 (2025-05-01)
 
 - Initial release.
+
+# Version 0.1.1 (2025-05-19)
+
+- Added dissapeared tests. (Qualitative x quantitative tests for 2 groups)
+- Fixed a bug where test did not run properly.
+- Changed package 'RVAideMemoire' -> 'DescTools' to match new versions.
+- Added test strength, for more clear and insightful results.
+- Added colors to a printed tests for a more clear result.